move allReduce to mllib
